+++ src/concrete.rs
@@ -1,0 +1,6 @@
+//! Example implementations for various data types.
+
+pub mod matrix;
+#[cfg(feature = "portgraph")]
+pub mod portgraph;
+pub mod string;
