+++ src/concrete/portgraph/predicate.rs
