+++ src/concrete/string/pattern.rs
