+++ src/concrete/string/predicate.rs
