+++ src/lib.rs
@@ -2,16 +2,13 @@
 #![warn(missing_docs)]
 
 pub mod automaton;
+pub mod concrete;
 pub mod constraint;
 pub mod constraint_tree;
 pub mod indexing;
 pub mod matcher;
-pub mod matrix;
 pub mod pattern;
-#[cfg(feature = "portgraph")]
-pub mod portgraph;
 pub mod predicate;
-pub mod string;
 pub mod utils;
 
 pub use constraint::{Constraint, DetHeuristic};
